Refactor ParallelStreams to use ForkJoinPool
The ParallelStreams class has been updated to use a ForkJoinPool, replacing the previous parallelStream. The reason for applying this change is to help handle complex CPU-intensive tasks more efficiently. More specifically, ForkJoinPool is a better fit when the workload is computation-heavy and takes a significant amount of time per item, which was causing potential fairness issues in the previous implementation.